requirements.txt
# BioResilient AI — Core Dependencies
# Generated from environment.yml for pip compatibility
# Updated for Python 3.12 compatibility

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Bioinformatics Core
biopython==1.84
numpy>=1.26.0

# Configuration
pyyaml==6.0.1

# HTTP & APIs
requests==2.31.0
aiohttp==3.9.1

# LLM (research assistant narrative)
anthropic>=0.18.0

# Parsing
lxml==4.9.3

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0

